refactor: Move pages/index.vue to app.vue to remove vue-router
